Project1/deliverable/project_1/pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.69.0",
    "beautifulsoup4>=4.14.2",
    "ddgs>=9.6.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "openai>=2.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
]
